reorganize cli command, update doc. Needed for future work with dependency visualizer
