+++ examples/README.md
@@ -5,7 +5,7 @@ Each example is already setup and initialize to use napi.
 ### To view the project on the UI
 
 ```
-napi ui
+napi split configure
 ```
 
 ### Or to split the APIs from the command line directly
