+++ packages/cli/src/cli/helpers/options.ts
@@ -1,0 +1,10 @@
+import { PositionalOptionsType } from "yargs";
+
+export const globalOptions = {
+  workdir: {
+    type: "string" as PositionalOptionsType,
+    default: process.cwd(),
+    alias: "wd",
+    description: "working directory",
+  },
+};
